server/requirements.txt
alembic==1.16.5
aniso8601==10.0.1
backports-datetime-fromisoformat==2.0.3
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
greenlet==3.2.4
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.0.1
PyJWT==2.10.1
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
six==1.17.0
SQLAlchemy==2.0.44
tomli==2.3.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==2.3.7
zipp==3.23.0
requests==2.31.0
python-dotenv==1.0.0
sendgrid==6.11.0
